Move module loaders from `kopf.reactor` to `kopf.utilities`
