+++ kopf/utilities/__init__.py
@@ -1,0 +1,5 @@
+"""
+General-purpose helpers not related to the framework itself
+(neither to the reactor nor to the engines nor to the structs),
+which are used to prepare and control the runtime environment.
+"""
