+++ tests/cli/test_help.py
@@ -16,7 +16,7 @@ def test_help_in_root(invoke, mocker):
 
 def test_help_in_subcommand(invoke, mocker):
     login = mocker.patch('kopf.config.login')
-    preload = mocker.patch('kopf.reactor.loading.preload')
+    preload = mocker.patch('kopf.utilities.loaders.preload')
     real_run = mocker.patch('kopf.reactor.queueing.run')
 
     result = invoke(['run', '--help'])
